Refactor: Repository Pattern 적용
