+++ app/src/main/java/com/example/simya/src/data/network/api/login/LoginRetrofitInterface.kt
@@ -1,7 +1,7 @@
-package com.example.simya.src.main.login.model
+package com.example.simya.src.data.network.api.login
 
-import com.example.simya.src.model.account.AccountDTO
-import com.example.simya.src.model.account.AccountResponse
+import com.example.simya.src.data.network.model.login.AccountDTO
+import com.example.simya.src.data.network.model.login.AccountResponse
 import retrofit2.Call
 import retrofit2.http.Body
 import retrofit2.http.Headers
