+++ app/src/main/java/com/example/simya/src/data/network/model/login/AccountDTO.kt
@@ -1,4 +1,4 @@
-package com.example.simya.src.model.account
+package com.example.simya.src.data.network.model.login
 
 import com.google.gson.annotations.SerializedName
