+++ app/src/main/java/com/example/simya/src/data/repository/login/LoginRepository.kt
@@ -1,9 +1,9 @@
-package com.example.simya.src.main.login.model
+package com.example.simya.src.data.repository.login
 
 import com.example.simya.config.BaseResponse
 import com.example.simya.src.data.network.model.login.AccountResponse
 
-interface LoginInterface {
+interface LoginRepository {
     fun onPostLoginSubmitSuccess(response: AccountResponse)
     fun onPostLoginSubmitFailure(response: BaseResponse)
     fun onPostLoginSubmitDisconnect(message: String)
