+++ app/src/main/java/com/example/simya/src/main/myPage/model/MyPageProfileRetrofitInterface.kt
@@ -1,7 +1,6 @@
 package com.example.simya.src.main.myPage.model
 
 import com.example.simya.config.BaseResponse
-import com.example.simya.src.model.account.AccountResponse
 import com.example.simya.src.model.profile.ProfileResponse
 import retrofit2.Call
 import retrofit2.http.*
