+++ app/src/main/java/com/example/simya/src/model/RetrofitService.kt
@@ -4,27 +4,11 @@ package com.example.simya.src.model
 import com.example.simya.config.BaseResponse
 import retrofit2.Call
 import retrofit2.http.*
-import com.example.simya.src.model.account.AccountDTO
-import com.example.simya.src.model.account.AccountResponse
-import com.example.simya.src.model.profile.ProfileResponse
-import com.example.simya.src.model.account.SignupDTO
-import com.example.simya.src.model.account.SignupResponse
-import com.example.simya.src.model.mypage.like.MyLikeStoryResponse
 import com.example.simya.src.model.mypage.review.MyModifyReviewResponse
 import com.example.simya.src.model.mypage.review.MyWriteReviewResponse
-import com.example.simya.src.model.profile.MyProfileResponse
 import com.example.simya.src.model.story.*
-import com.example.simya.src.model.story.create.CreateStoryDTO
-import com.example.simya.src.model.story.create.CreateStoryResponse
-import com.example.simya.src.model.story.inquiry.InquiryStoryDetailResponse
-import com.example.simya.src.model.story.load.LoadAllStoryResponse
-import com.example.simya.src.model.story.load.LoadMyStoryResponse
-import com.example.simya.src.model.story.open.OpenStoryDTO
-import com.example.simya.src.model.story.open.OpenStoryResponse
-import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.Headers
-import retrofit2.http.POST
 
 
 interface RetrofitService {
